Alertmanager integration test:  refactored and implemented fillInput() fix
