+++ frontend/packages/console-shared/src/test-utils/utils.ts
@@ -84,6 +84,22 @@ export async function click(elem: any, timeout?: number) {
   await elem.click();
 }
 
+export async function fillInput(elem: any, value: string) {
+  // Sometimes there seems to be an issue with clear() method not clearing the input
+  let attempts = 3;
+  do {
+    --attempts;
+    if (attempts < 0) {
+      throw Error(`Failed to fill input with value: '${value}'.`);
+    }
+    await browser.wait(until.and(until.presenceOf(elem), until.elementToBeClickable(elem)));
+    // TODO: line below can be removed when pf4 tables in use.
+    await elem.click();
+    await elem.clear();
+    await elem.sendKeys(value);
+  } while ((await elem.getAttribute('value')) !== value);
+}
+
 async function selectDropdownOptionByLocator(dropdownId: string, optionLocator: By) {
   await click($(dropdownId));
   await browser.wait(until.presenceOf(element(optionLocator)));
