+++ frontend/packages/kubevirt-plugin/integration-tests/tests/dialogs/cloneVirtualMachineDialog.ts
@@ -1,6 +1,6 @@
 import { browser, ExpectedConditions as until } from 'protractor';
-import { click } from '@console/shared/src/test-utils/utils';
-import { fillInput, selectOptionByText } from '../utils/utils';
+import { click, fillInput } from '@console/shared/src/test-utils/utils';
+import { selectOptionByText } from '../utils/utils';
 import { PAGE_LOAD_TIMEOUT_SECS } from '../utils/consts';
 import * as cloneDialogView from '../../views/dialogs/cloneVirtualMachineDialog.view';
